requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning & LLM Libraries
torch==2.0.1
transformers==4.32.1
datasets==2.14.0
tokenizers==0.13.3

# Model Development & Evaluation
evaluate==0.4.0
rouge-score==0.1.2
bert-score==0.3.13

# Visualization & Plotting
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework & Deployment
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database & Caching
sqlalchemy==2.0.20
redis==4.6.0

# Monitoring & Logging
python-json-logger==2.0.7
wandb==0.15.8

# Testing & Code Quality
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
isort==5.12.0

# ML Workflow & Experimentation
mlflow==2.7.0
pydantic==2.2.0

# Utilities
tqdm==4.65.0
requests==2.31.0
pyyaml==6.0.1
click==8.1.7
